requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
et-xmlfile==1.1.0
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.3
opencv-python==4.10.0.84
openpyxl==3.1.5
pandas==2.2.3
pygame==2.6.1
python-dateutil==2.9.0.post0
pytz==2024.2
packaging==24.2
pillow==11.0.0
pytesseract==0.3.13
requests==2.32.3
six==1.16.0
urllib3==2.2.3
Werkzeug==3.1.3
